Allow FlexBridge to run independent of FLEx
This change makes FlexBridge recognize a command line argument -runStandAlone.
If this is present it puts up status message boxes at all the points where
it would normally communicate with FLEx.

This allows the program to be directly debugged from Visual Studio.
Even more important, it can be debugged with MonoDevelop, which does
not support attaching a (useful) debugger to a running process.

Note: although this is mainly for the benefit of Linux, I wish it was
directly off develop rather than based on my other recent change. But it was
too hard to figure out how to do it on Linux. TortoiseHg there doesn't have
a stash!