Update download_dependencies scripts
This is mainly to pick up an extra .config file that's needed for
automatic dependency generation in the Linux package. However, it's
likely this is needed anyway if the GUI parts of FlexBridge are
to work at all.

(cherry picked from commit f758ff3b1f6c1adbafee1da7b5a310554602a76e)