Fix environ for fw geckofx upgrade (#114)
* Fix environ for fw geckofx upgrade

Set environ-xulrunner libraries to link to Geckofx45 which is
installed with Fieldworks. When running in Debug the libraries
are found in Fieldworks Debug output.

* update environ fix

Works with both install and debug. Moved below initial condition
that sets RUNMODE dependent on mono installation. There are also
changes to fieldworks environ due to the shared environ.
(see lib/common/setup-user.sh).