Fix problem when running unit tests
The unit tests of LfMerge used a directory that had the word
"Other" in the name. That caused GetMetaDataCacheOffsetFolder to
return a wrong directory which confused GetAdjustedCustomPropDirName
because it couldn't find the expected subdirectory. This change fixes
the problem by ensuring that there is a directory named exactly like
the expected one (instead of containing the string somewhere in the
path). While this doesn't solve the problem completely (there might
be a directory in the path that is named like the searched for
directory, but doesn't contain the expected information), it at
least improves things and makes the tests pass.