Fix DateModified if both changed different parts of entry
This fixes a problem when doing S/R in LanguageForge. LF relies on
the DateModified time stamp to know which entries it needs to
process. If an entry got modified in FLEx and the same entry but
different field got modified in LF, the S/R merged the changes but
set the DateModified field to the newer date (which was from LF).
This caused LF not to re-process the entry because it already
processed the entry with that timestamp. This caused the modified
data from FLEx to not show up in LF (https://trello.com/c/UoyXj4jP).

This change fixes this problem by setting DateModified to the
current time iff we and they change different fields of the same
entry.