Fix fixutf8
The fixutf8 extension appears to start off enabled with a relative
path. This doesn't work on Linux, and just adding a second line with
an absolute path doesn't fix it. So comment out any existing fixutf8
extension lines so we only have the one line with the absolute path.
Fixes LT-18796.