Ignore problematic DllImport references
MAPI32.DLL was Windows-only in MAPIHelper.cs before #if !MONO was
removed.
__Internal probably should have been ignored by dh_clideps without us
having to say so. This is in Platform.cs.
This fixes a Linux packaging error.

Both Platform.cs and MAPIHelper.cs are in libpalaso, and this problem
was introduced to flexbridge packaging when we switched to use the
master branch of libpalaso.