Fix handling of empty statusGuid
If we have an empty statusGuid but have a comment on both LD and LF
it means that we did a s/r before we introduced the statusGuid. Any
status change is assumed to have happened on LD, otherwise the s/r
would have set the statusGuid.

This change also switches the expected and actual values in the
calls to AssertThatXmlIn.