Fix bug when annotations contained no messages (#220)
* Fix bug when annotations contained no messages

Some annotations in real user data have NO <message> elements inside
them. This is not *supposed* to happen, but it does, so we need to deal
with it. This is the best way to do so.

* Make unit tests expect UTC

Now that the Chorus timestamp fix has been merged into flexbridge, the
unit tests' expectations need to be tweaked accordingly. Simplest way
to do it is to have the ReproducibleDateTimeProvider provide UTC dates.