Fix build (hopefully)
This change basically reverts commit e7177d8 which broke the build
on Linux. The NUnitTeamCity task causes a crash in Mono. Since the
reason why that change in October was necessary got fixed, we can
revert back to the old way. The SIL.BuildTasks assembly now ignores
warnings (I think).