Fix package build
Don't try to restore the packages during a binary package build.
This will fail because the current version of nuget doesn't work
with the mono version we're using during the package build. Instead
we download and install the packages when creating the source package.
This runs on the build agent (instead of the chroot), so we have a
working mono version available there.