Fix crash if we don't have a .ChorusNotes file (LF-199)
If we don't have a .ChorusNotes file we don't have to create one in
LfMergeBridge because that is done in Chorus. Just creating a text file
caused a crash when Chorus tried to read the file as XML.

The .ChorusNotes file won't exist if the user sends an empty project
(without any entries) in FLEx and we S/R to LF.

This change also adds a unit test for the case when the .ChorusNotes file
is missing.