rm environ
- We do not appear to use FB `environ` anywhere.
- FW's environ is inherited by FB when FW launches FB.
- It seems like fileslike `fieldworks-chorushub` would need settings
from `environ` to work. But they aren't actually _using_ `environ` at
present. They use `lib/flexbridge/run-app` and `environ-xulrunner` but
not `environ`. The `run-app` script has a lot of environment settings
in it, tho. However, `run-app` does look like it's falling back on
mono 4 not mono 5, so if this is working it may actually be using the
system mono.
- Remove `environ` so as not to be confusing.