launch: Optionally listen for debugger
- Usage: Set CHORUSDEBUGGING=true before launching.
- FB_MONO_ARGS and CHORUS_MONO_ARGS will allow other custom
commandline arguments to be externally set if helpful.
- Use non-default port 55502 for FlexBridge, and 55503 for
ChorusMerge, to not collide with others.
- Note that CHORUSDEBUGGING is also used by Chorus to print hg
debugging output. Setting in launch.json to set it for Chorus when
launching from launch.json, which should result both in hg debugging
output as well as listening for a debugger when ChorusMerge is later
run.