Obtain command can now clone a specific project
When given a -uri argument, the Obtain command will open the GetCloneFromInternetDialog and clone a new repo from the URI. I opted to expand this command over creating a new one, as it seemed a less intrusive and breaking change.

The URI is stripped of everything but its scheme, hostname, and project name (first segment). Further, a simple dialog box is shown on launch, confirming the download. These features help with security.

The operation returns a CloneResult, just like the "any project" operation, and will presumably proceed to open the new project.

The current design requires an environment variable to be set containing the JWT password. This allows us to avoid passing it through CMD to this process. It's recommended to instantiate the env var with scope limited to the process.

The JWT is coupled with the "bearer" username and therefore required. Other user/pass combinations are not supported.