Add support for opening pre-existing projects with custom names
If a -repositoryIdentifier (aka commit) is supplied from FLEx, we now look through the existing projects for a matching ID. If we find a match, progress through the clone as normal, indicating that a clone happened.

In coordination with this change, the FLEX obtain strategy now checks for previously existing .fwdata files in the indicated cloneLocation. If it finds such, it does not attempt to "put humpty together again."

Finally, in this case, it will proceed to "tell FLEx about it" in the respective method.