Expose fwdata dis/reassembly methods in LfMergeBridge (#398)
* Expose PutHumptyTogetherAgain method

Will be used for LexBox to be able to generate .fwdata files as needed

* Switch names of fwdata file APIs

The Humpty names are fun and cleverly silly, but when we expose them
to outside code in LfMerge, it might be nicer to have names that
communicates their purpose a little more clearly. We've settled on
DisassembleFwdataFile and ReassembleFwdataFile.