Allow passing comments to/from LfMerge w/o JSON
We use the ConditionalWeakTable class, designed to allow compilers to
attach metadata to objects, as an extra input/output mechanism that
doesn't require serializing the comments and replies to JSON strings.
This will save quite a lot of RAM when doing a Send/Receive of a project
with lots and lots of comments.

This implements the LfMergeBridge side of the process; a corresponding
change will be needed in LfMerge.