Be less strict about FixFwData name and location (#410)
On Linux, dotnet compiles binaries to filenames without .exe in them, so
we should handle that scenario gracefully. Also, it's common to be
running from a different directory than the one where FixFwData.exe is
found (e.g., `dotnet run` in the project that calls LfMergeBridge), so
we should also check the location of LfMergeBridge.dll since FixFwData
will usually be installed alongside LfMergeBridge.dll.